+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Added `WithSessionPoolNodeLimit` option for YDB.Open. It affects `table` and `query` internal session pools 
+and limits the number of sessions per node. Changes logic of session creation for cases where client uses `ydb.WithPreferredNodeID`.
+
 ## v3.125.1
 * Renamed `ydb_go_sdk_ydb_table_pool_node_hint_miss` and `ydb_go_sdk_ydb_query_pool_node_hint_miss` metrics to 
 `ydb_go_sdk_ydb_table_pool_node_hint` and `ydb_go_sdk_ydb_query_pool_node_hint`. Added `hit` label for them to